[OPP-1415] opprette overstyrt kodeverk for oppgave
Vi hardkoder de delene av kodeverket som ikke stemmer overens med defaulten for frist(=2 dager) og prioriteter(høy, normal, lav) og bruker dette til å overskrive verdier der det trengs.
